+++ apis/cluster/postgresql/v1alpha1/zz_postgresqlcluster_types.go
@@ -272,7 +272,7 @@ type PostgresqlClusterInitParameters struct {
 	// The total number of instances in the cluster (one master and n-1 standbys)
 	Instances *float64 `json:"instances,omitempty" tf:"instances,omitempty"`
 
-	// [string] The physical location where the cluster will be created. This will be where all of your instances live. Property cannot be modified after datacenter creation. Available locations: de/fra, us/las, us/ewr, de/txl, gb/lhr, gb/bhx, es/vit, fr/par, us/mci, de/fra/2. This attribute is immutable (disallowed in update requests).
+	// [string] The physical location where the cluster will be created. This will be where all of your instances live. Property cannot be modified after datacenter creation. Available locations: de/fra, us/las, us/ewr, de/txl, gb/lhr, gb/bhx, es/vit, fr/par, us/mci, de/fra/1, de/fra/2. This attribute is immutable (disallowed in update requests).
 	// The physical location where the cluster will be created. This will be where all of your instances live. Property cannot be modified after datacenter creation (disallowed in update requests)
 	// +crossplane:generate:reference:type=github.com/ionos-cloud/provider-upjet-ionoscloud/apis/cluster/compute/v1alpha1.Datacenter
 	// +crossplane:generate:reference:extractor=github.com/ionos-cloud/provider-upjet-ionoscloud/config/common.DatacenterLocation()
@@ -306,7 +306,7 @@ type PostgresqlClusterInitParameters struct {
 	// The storage type used in your cluster.
 	StorageType *string `json:"storageType,omitempty" tf:"storage_type,omitempty"`
 
-	// [string] Represents different modes of replication. Can have one of the following values: ASYNCHRONOUS, STRICTLY_SYNCHRONOUS. The SYNCHRONOUS value has been deprecated for create requests.
+	// [string] Represents different modes of replication. Can have one of the following values: ASYNCHRONOUS, STRICTLY_SYNCHRONOUS.
 	// Represents different modes of replication.
 	SynchronizationMode *string `json:"synchronizationMode,omitempty" tf:"synchronization_mode,omitempty"`
 }
@@ -355,7 +355,7 @@ type PostgresqlClusterObservation struct {
 	// The total number of instances in the cluster (one master and n-1 standbys)
 	Instances *float64 `json:"instances,omitempty" tf:"instances,omitempty"`
 
-	// [string] The physical location where the cluster will be created. This will be where all of your instances live. Property cannot be modified after datacenter creation. Available locations: de/fra, us/las, us/ewr, de/txl, gb/lhr, gb/bhx, es/vit, fr/par, us/mci, de/fra/2. This attribute is immutable (disallowed in update requests).
+	// [string] The physical location where the cluster will be created. This will be where all of your instances live. Property cannot be modified after datacenter creation. Available locations: de/fra, us/las, us/ewr, de/txl, gb/lhr, gb/bhx, es/vit, fr/par, us/mci, de/fra/1, de/fra/2. This attribute is immutable (disallowed in update requests).
 	// The physical location where the cluster will be created. This will be where all of your instances live. Property cannot be modified after datacenter creation (disallowed in update requests)
 	Location *string `json:"location,omitempty" tf:"location,omitempty"`
 
@@ -379,7 +379,7 @@ type PostgresqlClusterObservation struct {
 	// The storage type used in your cluster.
 	StorageType *string `json:"storageType,omitempty" tf:"storage_type,omitempty"`
 
-	// [string] Represents different modes of replication. Can have one of the following values: ASYNCHRONOUS, STRICTLY_SYNCHRONOUS. The SYNCHRONOUS value has been deprecated for create requests.
+	// [string] Represents different modes of replication. Can have one of the following values: ASYNCHRONOUS, STRICTLY_SYNCHRONOUS.
 	// Represents different modes of replication.
 	SynchronizationMode *string `json:"synchronizationMode,omitempty" tf:"synchronization_mode,omitempty"`
 }
@@ -431,7 +431,7 @@ type PostgresqlClusterParameters struct {
 	// +kubebuilder:validation:Optional
 	Instances *float64 `json:"instances,omitempty" tf:"instances,omitempty"`
 
-	// [string] The physical location where the cluster will be created. This will be where all of your instances live. Property cannot be modified after datacenter creation. Available locations: de/fra, us/las, us/ewr, de/txl, gb/lhr, gb/bhx, es/vit, fr/par, us/mci, de/fra/2. This attribute is immutable (disallowed in update requests).
+	// [string] The physical location where the cluster will be created. This will be where all of your instances live. Property cannot be modified after datacenter creation. Available locations: de/fra, us/las, us/ewr, de/txl, gb/lhr, gb/bhx, es/vit, fr/par, us/mci, de/fra/1, de/fra/2. This attribute is immutable (disallowed in update requests).
 	// The physical location where the cluster will be created. This will be where all of your instances live. Property cannot be modified after datacenter creation (disallowed in update requests)
 	// +crossplane:generate:reference:type=github.com/ionos-cloud/provider-upjet-ionoscloud/apis/cluster/compute/v1alpha1.Datacenter
 	// +crossplane:generate:reference:extractor=github.com/ionos-cloud/provider-upjet-ionoscloud/config/common.DatacenterLocation()
@@ -471,7 +471,7 @@ type PostgresqlClusterParameters struct {
 	// +kubebuilder:validation:Optional
 	StorageType *string `json:"storageType,omitempty" tf:"storage_type,omitempty"`
 
-	// [string] Represents different modes of replication. Can have one of the following values: ASYNCHRONOUS, STRICTLY_SYNCHRONOUS. The SYNCHRONOUS value has been deprecated for create requests.
+	// [string] Represents different modes of replication. Can have one of the following values: ASYNCHRONOUS, STRICTLY_SYNCHRONOUS.
 	// Represents different modes of replication.
 	// +kubebuilder:validation:Optional
 	SynchronizationMode *string `json:"synchronizationMode,omitempty" tf:"synchronization_mode,omitempty"`