+++ apis/namespaced/zz_register.go
@@ -19,9 +19,11 @@ import (
 	v1alpha1dbaas "github.com/ionos-cloud/provider-upjet-ionoscloud/apis/namespaced/dbaas/v1alpha1"
 	v1alpha1dns "github.com/ionos-cloud/provider-upjet-ionoscloud/apis/namespaced/dns/v1alpha1"
 	v1alpha1inmemorydb "github.com/ionos-cloud/provider-upjet-ionoscloud/apis/namespaced/inmemorydb/v1alpha1"
+	v1alpha1inmemorydbv2 "github.com/ionos-cloud/provider-upjet-ionoscloud/apis/namespaced/inmemorydbv2/v1alpha1"
 	v1alpha1k8s "github.com/ionos-cloud/provider-upjet-ionoscloud/apis/namespaced/k8s/v1alpha1"
 	v1alpha1kafka "github.com/ionos-cloud/provider-upjet-ionoscloud/apis/namespaced/kafka/v1alpha1"
 	v1alpha1log "github.com/ionos-cloud/provider-upjet-ionoscloud/apis/namespaced/log/v1alpha1"
+	v1alpha1mariadbv2 "github.com/ionos-cloud/provider-upjet-ionoscloud/apis/namespaced/mariadbv2/v1alpha1"
 	v1alpha1mongodb "github.com/ionos-cloud/provider-upjet-ionoscloud/apis/namespaced/mongodb/v1alpha1"
 	v1alpha1natgateway "github.com/ionos-cloud/provider-upjet-ionoscloud/apis/namespaced/natgateway/v1alpha1"
 	v1alpha1nfs "github.com/ionos-cloud/provider-upjet-ionoscloud/apis/namespaced/nfs/v1alpha1"
@@ -47,9 +49,11 @@ func init() {
 		v1alpha1dbaas.SchemeBuilder.AddToScheme,
 		v1alpha1dns.SchemeBuilder.AddToScheme,
 		v1alpha1inmemorydb.SchemeBuilder.AddToScheme,
+		v1alpha1inmemorydbv2.SchemeBuilder.AddToScheme,
 		v1alpha1k8s.SchemeBuilder.AddToScheme,
 		v1alpha1kafka.SchemeBuilder.AddToScheme,
 		v1alpha1log.SchemeBuilder.AddToScheme,
+		v1alpha1mariadbv2.SchemeBuilder.AddToScheme,
 		v1alpha1mongodb.SchemeBuilder.AddToScheme,
 		v1alpha1natgateway.SchemeBuilder.AddToScheme,
 		v1alpha1nfs.SchemeBuilder.AddToScheme,